added cancel button with some logic / still needs work
