+++ client/src/components/book/Book.js
@@ -1,5 +1,4 @@
-import React, { useEffect, useState } from "react";
-import { Card, CardBody } from "reactstrap";
+import Card from "react-bootstrap/Card";
 import { Link } from "react-router-dom";
 
 const Book = ({ book }) => {
@@ -25,7 +24,9 @@ const Book = ({ book }) => {
         bool = false;
       }
       else bool = true;
-    })
+      return bool
+    }
+    )
     return (
       bool ?
         <>
@@ -44,8 +45,8 @@ const Book = ({ book }) => {
   return (
     <>
       <Card>
-        <CardBody>
-          <img src={ thumbnail() } alt={ `Image of ${ book?.title }` } />
+        <Card.Body>
+          <img src={ thumbnail() } alt={ `${ book?.title }` } />
           <h3>{ book?.title }</h3>
           <h4>Author(s):</h4>
           <Author />
@@ -54,7 +55,7 @@ const Book = ({ book }) => {
               Details
             </button>
           </Link>
-        </CardBody>
+        </Card.Body>
       </Card>
     </>
   )
