+++ client/src/components/loan/Loan.js
@@ -1,11 +1,14 @@
 import React, { useEffect, useState } from "react";
 import Card from "react-bootstrap/Card";
 import Button from "react-bootstrap/Button"
-import { Link } from "react-router-dom";
 import { dateFixer } from "../../modules/helpers";
 import { updateLoanStatus } from "../../modules/loanManager";
+import Modal from 'react-bootstrap/Modal';
 
 const Loan = ({ loan, fetchLoans }) => {
+  const [ show, setShow ] = useState(false);
+  const [ loanEdit, setLoanEdit ] = useState({});
+  const [ isLoading, setIsLoading ] = useState(false)
   const [ status, setStatus ] = useState(
     {
       id: loan.id,
@@ -15,6 +18,25 @@ const Loan = ({ loan, fetchLoans }) => {
         status: ''
       }
     });
+
+  // When called, closes the Modal
+  const handleClose = () => {
+    setShow(false)
+  };
+
+  // Executes the modal
+  const handleShow = () => setShow(true);
+
+  // Sets the entry to be edited
+  const handleControlledInputChange = (e) => {
+    const newLoan = { ...loan };
+    let selectedVal = e.target.value;
+    newLoan.dueDateUnix = Date.parse(selectedVal) / 1000;
+    newLoan.ownerId = loan.owner.id;
+    newLoan.borrowerId = loan.borrower.id
+    setLoanEdit(newLoan)
+  }
+
   const requestDate = dateFixer(loan.requestDate)
   const [ currentStatus, setCurrentStatus ] = useState("");
 
@@ -43,16 +65,17 @@ const Loan = ({ loan, fetchLoans }) => {
   }, [ loan ])
 
   const handleLoanApprove = () => {
-    status.loanStatus.status = 'IsApproved';
-    updateLoanStatus(status).then(fetchLoans)
+    loanEdit.loanStatus.status = 'IsApproved';
+    updateLoanStatus(loanEdit).then(() => {
+      handleClose()
+      fetchLoans()
+    })
   }
 
   const handleLoanDeny = () => {
     status.loanStatus.status = 'IsDenied';
-    console.log('status', status)
     updateLoanStatus(status).then(fetchLoans)
   }
-
 
   return (
     <>
@@ -62,7 +85,7 @@ const Loan = ({ loan, fetchLoans }) => {
           <h4>Status: { currentStatus }</h4>
           <h4>Requested On: { requestDate }</h4>
 
-          <Button onClick={ () => handleLoanApprove() }>
+          <Button onClick={ () => handleShow() }>
             Approve
           </Button>
 
@@ -71,6 +94,48 @@ const Loan = ({ loan, fetchLoans }) => {
           </Button>
         </Card.Body>
       </Card>
+
+      <Modal show={ show } onHide={ handleClose }>
+        <Modal.Header closeButton>
+          <Modal.Title>Submit Loan Request</Modal.Title>
+        </Modal.Header>
+        <Modal.Body>
+          <form>
+            <fieldset>
+              <div className='habit-form__group'>
+                <label htmlFor='returnDate'>Return Date:</label>
+                <input
+                  type='date'
+                  required
+                  id='returnDate'
+                  onChange={ handleControlledInputChange }
+                  autoFocus
+                  className='form-control'
+                  defaultValue={ loan.date }
+                />
+              </div>
+            </fieldset>
+          </form>
+        </Modal.Body>
+        <Modal.Footer>
+          <div className='button-container'>
+            <div className='button-container__save'>
+              <Button
+                variant="secondary"
+                onClick={ handleClose }>
+                Close
+              </Button>
+              <Button
+                variant="primary"
+                onClick={ handleLoanApprove }
+                disabled={ isLoading }
+              >
+                Save Changes
+              </Button>
+            </div>
+          </div>
+        </Modal.Footer>
+      </Modal>
     </>
   )
 };
